+++ libs/react-components/src/tedi/components/heading-with-icon/heading-with-icon.stories.tsx
@@ -5,7 +5,7 @@ import HeadingWithIcon, { HeadingWithIconProps } from './heading-with-icon';
 
 /**
  * [Figma ↗](https://www.figma.com/file/jWiRIXhHRxwVdMSimKX2FF/TEDI-Design-System-(draft)?type=design&node-id=2137-19827&mode=dev)<br/>
- * [Zeroheight ↗](https://zeroheight.com/1ee8444b7/p/94147f-heading-with-icon/b/23093e)
+ * [Zeroheight ↗](https://tedi.tehik.ee/1ee8444b7/p/94147f-heading-with-icon)
  */
 
 export default {