+++ libs/react-components/src/tedi/index.ts
@@ -14,6 +14,7 @@ export * from './components/separator/separator';
 export * from './components/buttons/closing-button/closing-button';
 export * from './components/print/print';
 export * from './components/buttons/button/button';
+export * from './components/alert/alert';
 export * from './components/stretch-content/stretch-content';
 export * from './components/navigation/link/link';